backend/test_gemini.py
import os
import google.generativeai as genai
from dotenv import load_dotenv
load_dotenv()
genai.configure(api_key=os.environ["GEMINI_API_KEY"])
print("Available models:")
for m in genai.list_models():
    print(m.name)
