backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
crewai
crewai-tools
litellm
chromadb>=0.5.0
langchain>=0.2.0
langchain-community
tavily-python
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
sentence-transformers>=2.7.0
